refactor(cli): extract parser crate
Move the hand-written ccusage CLI parser, help spec, and parser snapshots into a new internal ccusage-cli workspace crate. This keeps parser tests focused and gives future parser experiments a narrow crate boundary without adding a runtime parser dependency.

Keep the main ccusage crate responsible for config discovery by passing ConfigContext through the new CliConfig trait. Move CLI help generation into the new crate build script and pre-render merged option sections at build time to avoid runtime option concatenation for common contextual help output.

The runtime crate now re-exports the extracted CLI types through its existing cli module and uses a local TerminalStyle helper where the extracted SharedArgs type can no longer satisfy the orphan rules.
