+++ rust/Cargo.toml
@@ -1,5 +1,10 @@
 [workspace]
-members = ["crates/ccusage", "crates/ccusage-terminal", "crates/ccusage-test-support"]
+members = [
+	"crates/ccusage",
+	"crates/ccusage-cli",
+	"crates/ccusage-terminal",
+	"crates/ccusage-test-support",
+]
 resolver = "2"
 
 [profile.release]
