+++ rust/crates/ccusage-cli/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ccusage-cli"
+version = "20.0.4"
+edition = "2021"
+build = "build.rs"
+
+[dependencies]
+
+[build-dependencies]
+serde_json = "1"
+
+[dev-dependencies]
+insta = { version = "1.47.2", features = ["json"] }
+serde_json = "1"
