+++ rust/crates/ccusage-cli/build.rs
@@ -1,0 +1,157 @@
+use std::{collections::BTreeMap, env, fs, path::PathBuf};
+
+use serde_json::{Map, Value};
+
+const CLI_HELP_JSON: &str = "src/cli-help.json";
+
+fn main() {
+    println!("cargo:rerun-if-changed={CLI_HELP_JSON}");
+    generate_cli_help_rs();
+}
+
+fn out_dir_path(file_name: &str) -> PathBuf {
+    PathBuf::from(env::var_os("OUT_DIR").expect("OUT_DIR is set by cargo")).join(file_name)
+}
+
+fn generate_cli_help_rs() {
+    let source = fs::read_to_string(CLI_HELP_JSON).expect("read CLI help spec");
+    let Value::Object(option_sets) =
+        serde_json::from_str::<Value>(&source).expect("parse CLI help spec")
+    else {
+        panic!("CLI help spec must be a JSON object");
+    };
+
+    let mut output = String::from("// @generated by build.rs\n\n");
+    let mut rendered_options = BTreeMap::new();
+    let mut option_sets = option_sets.iter().collect::<Vec<_>>();
+    option_sets.sort_by(|(left, _), (right, _)| left.cmp(right));
+    for (function_name, options) in option_sets {
+        let rendered = render_options(options);
+        rendered_options.insert(function_name.to_string(), rendered.clone());
+        output.push_str(&format!(
+            "#[allow(dead_code)]\nfn {function_name}() -> &'static str {{\n    {}\n}}\n\n",
+            rust_string_literal(&rendered)
+        ));
+    }
+    for (function_name, parts) in [
+        (
+            "all_agent_session_options",
+            &["all_agent_options", "session_options"][..],
+        ),
+        (
+            "claude_daily_options",
+            &["shared_claude_options", "daily_options"][..],
+        ),
+        (
+            "claude_weekly_options",
+            &["shared_claude_options", "weekly_options"][..],
+        ),
+        (
+            "claude_session_options",
+            &["shared_claude_options", "session_options"][..],
+        ),
+        (
+            "blocks_combined_options",
+            &["shared_claude_options", "blocks_options"][..],
+        ),
+        ("pi_combined_options", &["agent_options", "pi_options"][..]),
+        (
+            "openclaw_combined_options",
+            &["agent_options", "openclaw_options"][..],
+        ),
+    ] {
+        let rendered = render_combined_options(&rendered_options, parts);
+        output.push_str(&format!(
+            "#[allow(dead_code)]\nfn {function_name}() -> &'static str {{\n    {}\n}}\n\n",
+            rust_string_literal(&rendered)
+        ));
+    }
+    fs::write(out_dir_path("cli-help.rs"), output).expect("write generated CLI help");
+}
+
+fn render_options(options: &Value) -> String {
+    let Value::Array(options) = options else {
+        panic!("CLI help option set must be an array");
+    };
+    let width = options
+        .iter()
+        .map(|option| {
+            let Value::Object(option) = option else {
+                panic!("CLI help option must be an object");
+            };
+            string_field(option, "flags").len()
+        })
+        .max()
+        .unwrap_or(36)
+        .max(36);
+    let mut lines = vec!["OPTIONS:".to_string()];
+    for option in options {
+        let Value::Object(option) = option else {
+            panic!("CLI help option must be an object");
+        };
+        let flags = string_field(option, "flags");
+        let description = string_field(option, "description");
+        let mut line = format!("  {flags:<width$} {description}");
+        let mut details = Vec::new();
+        if let Some(default) = optional_string_field(option, "default") {
+            details.push(format!("default: {default}"));
+        }
+        if let Some(choices) = option.get("choices") {
+            let Value::Array(choices) = choices else {
+                panic!("CLI help choices must be an array");
+            };
+            if !choices.is_empty() {
+                details.push(format!(
+                    "choices: {}",
+                    choices
+                        .iter()
+                        .map(|choice| {
+                            choice
+                                .as_str()
+                                .expect("CLI help choice must be a string")
+                                .to_string()
+                        })
+                        .collect::<Vec<_>>()
+                        .join(" | ")
+                ));
+            }
+        }
+        if !details.is_empty() {
+            line.push_str(&format!(" ({})", details.join(", ")));
+        }
+        lines.push(line);
+    }
+    lines.join("\n")
+}
+
+fn render_combined_options(options: &BTreeMap<String, String>, parts: &[&str]) -> String {
+    let option_lines = parts
+        .iter()
+        .flat_map(|part| {
+            options
+                .get(*part)
+                .unwrap_or_else(|| panic!("missing generated option set {part}"))
+                .lines()
+                .skip(1)
+        })
+        .collect::<Vec<_>>()
+        .join("\n");
+    format!("OPTIONS:\n{option_lines}")
+}
+
+fn string_field(object: &Map<String, Value>, field: &str) -> String {
+    optional_string_field(object, field).unwrap_or_else(|| panic!("missing CLI help field {field}"))
+}
+
+fn optional_string_field(object: &Map<String, Value>, field: &str) -> Option<String> {
+    object.get(field).map(|value| {
+        value
+            .as_str()
+            .unwrap_or_else(|| panic!("CLI help field {field} must be a string"))
+            .to_string()
+    })
+}
+
+fn rust_string_literal(value: &str) -> String {
+    serde_json::to_string(value).expect("encode Rust string literal")
+}
