+++ rust/crates/ccusage/Cargo.toml
@@ -13,6 +13,7 @@ name = "generate-config-schema"
 path = "src/bin/generate_config_schema.rs"
 
 [dependencies]
+ccusage-cli = { path = "../ccusage-cli" }
 ccusage-terminal = { path = "../ccusage-terminal" }
 jiff = { version = "0.2.24", default-features = false, features = [
 	"std",
