+++ rust/crates/ccusage/build.rs
@@ -2,7 +2,6 @@ use std::{env, fs, path::PathBuf};
 
 use serde_json::{Map, Value};
 
-const CLI_HELP_JSON: &str = "src/cli-help.json";
 const FLAKE_LOCK_JSON: &str = "../../../flake.lock";
 const LITELLM_PRICING_JSON: &str = "model_prices_and_context_window.json";
 const OUT_PRICING_JSON: &str = "litellm-pricing.json";
@@ -10,10 +9,7 @@ const PRICING_JSON_PATH_ENV: &str = "CCUSAGE_PRICING_JSON_PATH";
 const PRICING_FETCH_TIMEOUT_SECONDS: u64 = 10;
 
 fn main() {
-    println!("cargo:rerun-if-changed={CLI_HELP_JSON}");
     println!("cargo:rerun-if-env-changed={PRICING_JSON_PATH_ENV}");
-
-    generate_cli_help_rs();
 
     let out_path = out_dir_path(OUT_PRICING_JSON);
     let pricing_json = if let Some(path) = env::var_os(PRICING_JSON_PATH_ENV) {
@@ -31,99 +27,6 @@ fn main() {
 
 fn out_dir_path(file_name: &str) -> PathBuf {
     PathBuf::from(env::var_os("OUT_DIR").expect("OUT_DIR is set by cargo")).join(file_name)
-}
-
-fn generate_cli_help_rs() {
-    let source = fs::read_to_string(CLI_HELP_JSON).expect("read CLI help spec");
-    let Value::Object(option_sets) =
-        serde_json::from_str::<Value>(&source).expect("parse CLI help spec")
-    else {
-        panic!("CLI help spec must be a JSON object");
-    };
-
-    let mut output = String::from("// @generated by build.rs\n\n");
-    let mut option_sets = option_sets.iter().collect::<Vec<_>>();
-    option_sets.sort_by(|(left, _), (right, _)| left.cmp(right));
-    for (function_name, options) in option_sets {
-        let rendered = render_options(options);
-        output.push_str(&format!(
-            "fn {function_name}() -> &'static str {{\n    {}\n}}\n\n",
-            rust_string_literal(&rendered)
-        ));
-    }
-    fs::write(out_dir_path("cli-help.rs"), output).expect("write generated CLI help");
-}
-
-fn render_options(options: &Value) -> String {
-    let Value::Array(options) = options else {
-        panic!("CLI help option set must be an array");
-    };
-    let width = options
-        .iter()
-        .map(|option| {
-            let Value::Object(option) = option else {
-                panic!("CLI help option must be an object");
-            };
-            string_field(option, "flags").len()
-        })
-        .max()
-        .unwrap_or(36)
-        .max(36);
-    let mut lines = vec!["OPTIONS:".to_string()];
-    for option in options {
-        let Value::Object(option) = option else {
-            panic!("CLI help option must be an object");
-        };
-        let flags = string_field(option, "flags");
-        let description = string_field(option, "description");
-        let mut line = format!("  {flags:<width$} {description}");
-        let mut details = Vec::new();
-        if let Some(default) = optional_string_field(option, "default") {
-            details.push(format!("default: {default}"));
-        }
-        if let Some(choices) = option.get("choices") {
-            let Value::Array(choices) = choices else {
-                panic!("CLI help choices must be an array");
-            };
-            if !choices.is_empty() {
-                details.push(format!(
-                    "choices: {}",
-                    choices
-                        .iter()
-                        .map(|choice| {
-                            choice
-                                .as_str()
-                                .expect("CLI help choice must be a string")
-                                .to_string()
-                        })
-                        .collect::<Vec<_>>()
-                        .join(" | ")
-                ));
-            }
-        }
-        if !details.is_empty() {
-            line.push_str(&format!(" ({})", details.join(", ")));
-        }
-        lines.push(line);
-    }
-    lines.join("\n")
-}
-
-fn string_field(object: &Map<String, Value>, field: &str) -> String {
-    optional_string_field(object, field).unwrap_or_else(|| panic!("missing CLI help field {field}"))
-}
-
-fn optional_string_field(object: &Map<String, Value>, field: &str) -> Option<String> {
-    object.get(field).map(|value| {
-        value
-            .as_str()
-            .unwrap_or_else(|| panic!("CLI help field {field} must be a string"))
-            .to_string()
-    })
-}
-
-fn rust_string_literal(value: &str) -> String {
-    serde_json::to_string(value).expect("encode Rust string literal")
 }
 
 fn fetch_pricing_json() -> std::io::Result<String> {
